+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Added ECS tool and run for pre-craft-4 cleanup
 -   Added craftcms/phpstan and run for pre-craft-4 cleanup
 -   Created `OrderValidator` service class and moved relevant code into it.
+-   Created `OrderMetadata` service class and moved relevant code into it.
 
 ## 1.2.0 - 2023-11-06